Add BleScanner class and integrate in main
Note: not tested in Unity, because the BleScanner class uses many custom types from the NimBLE library, which cannot be compiled on the native platform.

The same goes for main.cpp.

Since this concerns only a few lines of code, this is acceptable for now.

Alternatively, you can write Unity tests for the embedded platform. However, this still cannot test everything and cannot be automated without significant commitment.